+++ lib/src/analyzers/unnecessary_nullable_analyzer/reporters/unnecessary_nullable_report_params.dart
@@ -1,0 +1,6 @@
+/// Represents additional unnecessary nullable reporter params.
+class UnnecessaryNullableReportParams {
+  final bool congratulate;
+
+  const UnnecessaryNullableReportParams({required this.congratulate});
+}