+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_expanded_as_spacer/examples/example.dart
@@ -1,0 +1,48 @@
+class MyWidget extends StatelessWidget {
+  Widget build(BuildContext context) => Column(children: [
+        Expanded(child: Container()), // LINT
+        Expanded(child: SizeBox()), // LINT
+        Expanded(child: foo()),
+        Expanded(child: SizeBox(child: Container())),
+        Expanded(child: Container(child: SizeBox())),
+        Spacer(),
+      ]);
+}
+
+Widget foo() => SizeBox();
+
+class Container extends Widget {
+  final Widget? child;
+
+  const Container({this.child});
+}
+
+class Expanded extends Widget {
+  final Widget? child;
+
+  const Expanded({this.child});
+}
+
+class SizeBox extends Widget {
+  final Widget? child;
+
+  const SizeBox({this.child});
+}
+
+class Spacer extends Widget {
+  const Spacer();
+}
+
+class Column extends Widget {
+  final List<Widget> children;
+
+  Column({required this.children});
+}
+
+class StatelessWidget extends Widget {}
+
+class Widget {
+  const Widget();
+}
+
+class BuildContext {}