+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_commenting_analyzer_ignores/examples/example.dart
@@ -15,4 +15,7 @@ void main() {
 
   // ignore: deprecated_member_use, long-method multiple same line ignore
   final linkedList = LinkedList();
+
+  // ignore: avoid-non-null-assertion, checked for non-null
+  final hashMap = HashMap();
 }